ai/gpt-sample/src/multiagent.py
from autogen import AutoGen
from langchain import LangChain

autogen = AutoGen()
langchain = LangChain()

# Create an agent in AutoGen
agent = autogen.create_agent()

# Use GPT-4 Turbo's language model to enhance the agent's capabilities
agent.set_language_model(chat.get_language_model())
