ai/gpt-sample/pyproject.toml
[tool.poetry]
name = "gpt-sample"
version = "0.0.0"
description = ""
authors = ["alstjr7375 <alstjr7375@daum.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
python-dotenv = "^1.0.0"
openai = "^1.3.6"
langchain = "^0.0.343"
pyautogen = "^0.2.0"

[tool.poetry.group.dev.dependencies]
pyright = "^1.1.338"
ruff = "^0.1.6"

[tool.pyright]
# https://microsoft.github.io/pyright/#/configuration?id=sample-pyprojecttoml-file
include = ["src"]
exclude = [
  "**/node_modules",
  "**/__pycache__",
  "**/.venv"
]
defineConstant = { DEBUG = true }

reportMissingImports = true
reportMissingTypeStubs = false
strictListInference = true
strictDictionaryInference = true
strictSetInference = true

pythonVersion = "3.11"

# https://docs.astral.sh/ruff/tutorial/#getting-started
[tool.ruff]
# Set the maximum line length to 100
line-length = 100
indent-width = 2

[tool.ruff.lint]
# Add the `line-too-long` rule to the enforced rule set. By default, Ruff omits rules that
# overlap with the use of a formatter, like Black, but we can override this behavior by
# explicitly adding the rule.
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
extend-select = ["E501"]

[tool.ruff.format]
skip-magic-trailing-comma = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
